did a bunch of setup stuff
