+++ www/app/admin.py
@@ -1,3 +1,3 @@
 from django.contrib import admin
 
-# Register your models here.
+# Register your domain here.
