+++ www/app/domain/models/__init__.py
@@ -1,0 +1,3 @@
+from .Account import Account
+from .Course import Course
+from .Lab import Lab
