+++ www/app/views.py
@@ -1,9 +1,10 @@
 from django.shortcuts import render
 from django.http import HttpResponse
-from www.app.domain import *
+from .domain.models import Account
 
 
 # Create your views here.
+
 def index(req):
     a = Account()
     return HttpResponse('Hello there')
